+++ src/main/java/io/github/astrapi69/jaxb/factory/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package provides factory classes
+ *
+ * @author Asterios Raptis
+ */
+package io.github.astrapi69.jaxb.factory;