+++ src/main/java/io/github/astrapi69/jaxb/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package provides classes for the transformation of objects to xml and back with jaxb
+ * 
+ * @author Asterios Raptis
+ */
+package io.github.astrapi69.jaxb;