+++ CHANGELOG.md
@@ -6,19 +6,21 @@ Version 1.2-SNAPSHOT
 
 ADDED:
 
+- new module-info.java file for module description
+- all required package-info.java files
 - new converter class that transforms a xml file to a java object
 - new extension class that transforms a xml file to a java object
 
 CHANGED:
 
-- update of gradle to new version 7.6-rc-1
+- update of gradle to new version 7.6-rc-2
 - update of gradle plugin dependency com.github.ben-manes.versions.gradle.plugin to new minor version 0.43.0
 - update of gradle-plugin dependency of 'io.freefair.gradle:lombok-plugin' in version 6.5.1
 - update of gradle-plugin dependency of 'com.diffplug.spotless:spotless-plugin-gradle' in version 6.11.0
-- update of dependency xml-api to new minor version 1.3
+- update of dependency xml-api to new minor version 1.4
 - update of dependency checksum-up to new minor version 2.2
 - update of dependency jaxb-runtime to new version 4.0.1
-- update of test dependency file-worker to new minor version 11.5
+- update of test dependency file-worker to new minor version 11.5.1
 - update of test dependency junit-jupiter-api to new minor version 5.9.1